requirements.txt
tensorflow==2.17.0
numpy==1.26.4
streamlit==1.38.0
pyyaml==6.0.2
scikit-learn==1.5.2
matplotlib==3.9.2